backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
joblib==1.3.2
Pillow==10.2.0
pytesseract==0.3.10
PyMuPDF==1.27.1
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
aiosqlite==0.19.0
httpx==0.26.0
email-validator==2.1.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
psutil==5.9.7
weasyprint==60.1

